+++ CHANGES.md
@@ -12,6 +12,7 @@ Changes:
 
 Fixes:
 - Modifies rendering of AhocorasickTokenizer parameter in API docs II
+- Fixes dropping of nominative reporter due to overlapping CitationTokens
 
 ## Current